Improve remaining character calc and selection
Refactor the code so that remaining characters is more robust, taking
into account if length argument is set or if the sum of the other
arguments are more than the default length.

This change also updates the character selection to extract a single
character at a time using sample, rather than selecting all characters
at once. This was problematic as letters could only be used once, now we
can have the same letter multiple times.