Use a wider mixture if character types
One problem with the code is that we select character types in order of
lowercase -> uppercase -> numbers -> special. This can often lead to
passwords being [a-z] heavy and not include many numbers or special
characters. When randomly assigning character counts we now limit the
possible total for each set by 60% of the remaining total. This provides
are more varied final set.