Refactor towards object oriented code
While previously using a more functional style of programming, by
passing arguments between methods, this commit alters the code to be
more object oriented in approach. Although we are still working with a
Ruby module, we are able to make use of private methods and instance
variables to share state between methods essentially creating a
singleton.

This allows a more concise implementation that is perhaps easier to
follow, with the tradeoff that methods are now using variables outside
of their scope. There is also a risk that this implementation may have
issues in a multi threaded environment as it uses instance variables. If
this is considered a risk we could refactor the code further to use a
class.