Refactor spec assertions so they are less brittle
Previously I was asserting the exact output for each test, this approach
proved to be too brittle as a minor modification to how the password
is constructed results in all tests needing to be updated.

The previous approach also relied too much on manual checks, in that a
human would need to manually count the number of each character type in
each test to determine if they were behaving correctly.

The previous approach also relied on the randomness of the password
generation to be locked, so that tests would behave the same on each
run. This results in tests that are less robust.

This commit refactors the tests so that I am instead checking the
composition of each result. This means I no longer need to manually
count characters in each test and the tests are more resilient.