Refactor to remove instance variables
The advantage of using a singleton means the module can make use of
private methods and instance variables so that arguments do not need to
be passed between methods.

The downsides to this though are that instance variables are not as
declarative and introduce a risk with thread safety as state is shared.

To mitigate any potential issues, this commit removes instance variables
and adds arguments to methods where needed.